Fix warnings when build with msvc 64bit

+++ core/src/oned/ODMultiUPCEANReader.cpp
@@ -346,7 +346,7 @@ Result MultiUPCEANReader::decodePattern(int rowNumber, const PatternView& row, s
 
 	auto ext = res.end;
 	PartialResult extRes;
-	if (ext.skipSymbol() && ext.skipSingle(begin.sum() * 3.5) &&
+	if (ext.skipSymbol() && ext.skipSingle(static_cast<int>(begin.sum() * 3.5)) &&
 		((expectExtension(5) && Extension(extRes, ext, 5)) || (expectExtension(2) && Extension(extRes, ext, 2)))) {
 
 		//TODO: extend position in include extension
